feat(utxo-lib): new function getTaprootOutputKey
To get x-only output key from taproot output script

TICKET: BG-66942
